%[BugId: 32]%
UVPDataTransfer now uses the new UVPDataAtom and UVPDataAtomHeader
classes.

It subscribes to messages from the integrator. It waits for a "Hello"
from the integrater and then publishes a message that starts the
integrator.

Did many optimizations in the receive method. The main timeconsumer in
that method is the std::map::operator[] method that is used by the
UVPDataSet class. Insertion in the map is very slow. Time for a new
internal datastructure?

removed the size() and getRow() methods. They were obsolete due to the
separation of the transfer (UVPADTATransferWP) and storage (UVPDataSet)
classes.