%[BugId: 64]%
Created buildMenuBar() and buildStatusBar() methods and moved
corresponding code from the constructor to these methods.

The class now has data members "itsInputType" and "itsInputFilename"
that store the type of input (None, DMI, MS, PVD) and the source file of
the input, if that file exists. This is an intermediate step in the
implementation of a slight change in the user interface regarding
opening files and loading baselines attached to a particular antenna.