%[BugId:95]%
meqcalibrater.g:
* getparms now takes an additional (but optional) argument to exclude parameters
which are selected in the parmpatterns. This is done using the excludepatterns
parameter.
* added new method getsolvedomain which returns the current solve domain. The
solve domain is updated after each call to nexttimeiteration.
* removed mepDB argument from the meqcalibrater constructor. The name of the MEP
database is now derived from the meqmodel parameter by appending the .MEP extention.

Update MeqCalibraterImpl implementation of the DO accordingly. Added comments to
the MeqCalibraterImpl.h file.