%[ER: 61]%
Removed exception class NotImplemented. Methods that are not (yet) implemented
are now simply removed from the implementation file. The advantage is that the
user is informed at link-time, instead of at run-time, when he/she uses a
feature that is not (yet) implemented. The header file should contain a "todo"
remark in the Doxygen documentation.