%[ER:95 ]%
The &timeout argument to select is filled with the remaining time of the timeout.
E.g. when the timeout is set to 5 seconds, but there is an event after 3 seconds,
the timeout return value is set to 2 seconds. We want a timeout of 10ms therefor
we need to reset timeout to 10ms on each new call to workProc (and select).