%[ER:70 ]%
Removed use of += within LOG4CPLUS condition. It doesn't work on older
automake versions.
Renamed LimitExec.c to LimitExec.cc because the gcc compiler could not
link with the c++ log4cplus library! This is a problem caused by the fact
that for any configured packages all libraries are unconditionally added to
LDADD. What we should really have is a LOG4CPLUS_LDADD which can then be
used as needed.