%[ER:97 ]%
Added possibility to have initializers on the default constructor. If
a member variable is of a class that does not have a default constructor
another constructor needs to be called in the initializers of the event class/struct. This can be done as follows:

event = {
	name = MYEVENT;
	dir = IN;
	param = {
		name = "test";
		type = "MyTypeWithOnlyConstructorThatTakesOneIntAsArgument";
		init = "1";
	};
};

This creates the following code for Myevent:

struct Myevent
{
	Myevent() : GCFEvent(MYEVENT), test(1)
	{
	}
};