%[ER: 65]%
Logical expressions (!, &&, and ||) no longer use logical expression nodes,
but their "ordinary" counterparts. It turned out that the join constraints
should *always* be ouput last, even for logical expression nodes. If you don't
do that, the generated query will be wrong, as the database engine will create
"unwanted" cartesian products for the different tables in the query. Hence,
the query will take a very long time to run and the result will be incorrect.