%[ER: 231]%
Merged the tests tConverterImpl and tConverterClient into tConverter. This
makes sense, because both programs shared a lot of source code and the output
of both programs should always be the same. Whether ConverterImpl or
ConverterClient is tested is now determined by the contents of tConverter.in;
this file is created by tConverter.sh. The tConverterServer program was
promoted to the program `amcserver' and moved into the src directory tree.