BugID: 360
Implemented configurable CEP output feature.
New RSPDriver.conf variables:

Set the payload size that is sent on the CEP output port.

- RSPDriver.OUTPUT_PAYLOAD_SIZE=736

For each board [0..23] the following three variables can be set
to control the source and destination IP-address and the
destination MAC address which will be set in the UDP/IP header.

- RSPDriver.SRC_IP_ADDR_0=10.1.0.1
- RSPDriver.DST_IP_ADDR_0=10.2.0.1
- RSPDriver.DST_MAC_ADDR_0=10:FA:00:00:00:01