BugID: 442
Added a simple kind of stress test. In its current form, the test program
makes 100 consecutive connections to the converter server and requests one
simple conversion. Its main purpose is to check for potential file descriptor
leaks in the server (which was present in an earlier version). However, the
program may be enhanced by making several concurrent connections.
Note: In order to trigger file descriptor leakage, the maximum number of open
files is set to 20 in the tConverterStress.run file.