BugID: 442
Needed to completely overhaul the interface for this, so it was more work than
I anticipated. However, the Converter interface is much cleaner now, so adding
a return status really had a positive side effect. Due to these interface
changes I'll have to turn this into a new major release (2.0).