BugID: 145 670
THROW is not redefined anymore in Exception.h (BugId #145), it will only be
defined if it wasn't defined previously. Note that LofarLogger.h, on the other
hand, will always (re)define THROW. Added a test to check this behaviour.

THROW now behaves as a valid statement (BugId #670), so you should always use
a semicolon to end a THROW "statement".