BugID: 697
Direction now inherits from Position.

Implementation is not perfect yet. I ran into troubles with the member
function `virtual Types type() const', because it has different return types
for Position and Direction. I (temporarily) solved this by making type()
return an int.

A better solution would be to promote the enum Types to a real class. This
would also solve the problem that you can now write

  Position pos(1, 1, Position::WGS84);  // Murphy

or worse

  Position pos(1,1, Direction::AZEL);   // Machiavelli

which will yield a position pos with lon=1, lat=1, height=1, type=ITRF, due to
implicit conversion of the enum to double. This is probably not what the
programmer intended!