BugID: 697
Added constructor to create a Position from a Coord3D. This was needed in
order to be able to assign the result of `Coord3D operator+(const Coord3D&,
const Coord3D&) to a Position.
Note: this is in fact a smell of bad design. I should implement Position and
Direction in terms of `has-a' Coord3D instead of `is-a' Coord3D.