BugID: 834
Removed AM_CONDITIONAL for HAVE_AIPSPP. It caused trouble when not being
defined (which occurs for example when AIPSPATH has not been set). It seems
that it is not used (anymore). At least, I could'nt find any occurrences of
HAVE_AIPSPP in any of the Makefile.am files in the LOFAR tree.