BugID: 921
Changed definition of CHECKTOOLCOMMAND. It now contains the command prefix
"$LOFAR_LIMITEXEC $MAXTIME". Note that this may cause some test programs to
fail when run through valgrind, because a valgrind run usually takes much
longer to complete. The solution would then be to increase MAXTIME for the
offending program in the associated *.sh file.