BugID: 686
Indirect package dependencies will now be derived by inspecting the contents
of the directory config/<pkg-name> in the install directory for all packages
mentioned in the current configure.in file.

The directory `config/<pkg-name>' contains the following files:
  pkgext         : list of all packages that the package <pkg-name> depends on.
  pkgextcppflags : preprocessor flags passed when <pkg-name> was built.
  pkgextcxxflags : C++ compiler flags passed when <pkg-name> was built.
  pkgextldflags  : linker flags passed when <pkg-name> was built.

The preprocessor flags (CPPFLAGS) and the linker flags (LDFLAGS) will be
concatenated, but the C++ compiler flags (CXXFLAGS) not. So, for example, when
one of the header files of LCS/Common includes a Boost header file, the
required -I<path> will be added to config/Common/pkgextcppflags. As a result,
when you include this specific Common header file, you do not have to specify
in your configure.in file anymore that you (indirectly) depend on Boost.