BugID: 983 Protocol extension: ------------------ Extended the protocol template with the ability of specifying errornumbers and errormessages. The generated errornumbers are combined with the (uniq) protocol- number so each errornumber will be uniq within MAC!
Syntax:
error = {
	id = RANGE;
	msg = "Illegal range specified";
};

This will result in an errornumber <PROTOCONAME>_RANGE_ERR.
To get the corresponding errormessage use: errorName(<PROTOCOLNAME>_RANGE_ERR)

Registration changes:
--------------------
Till now the registration of the protocols was a function of a GCF-Task. Therefore
it was not possible to print the names of the events in a 'plain' FSM. This
code change also introduces a new 'registerProtocol' function;

registerProtocol (unsigned short protID, struct protocolStrings& protDef)

The protocolStrings structure is defined in GCF/TM/GCF_Protocols.h and contains
pointers to the errormessages AND the signalnames. The new autogen template
generates a <PROTOCOLNAME>_STRINGS variable that can be used in the new
registerProtocol functions.
To print the signalname use: signalName(event) in stead of the 'old'
evtstr(event) function.

NOTE: the old GCFTask related functions registerProtocol and evtstr are still
available for backwards compatibility but they will become obsolete in the
(near) future.

For more information see the tProtocol.cc and testprotocol.prot file in the
test directory of GCF/TM.