BugID: 992
Using a bool to indicate case (in)sensitive key comparison was not such a good
choice. First, it causes trouble with the trivial conversion of const char* to
bool, requiring an extra ParameterSet constructor taking a const char*.
Second, it is not very verbose; i.e., what does it mean when I construct a
ParameterSet(true)?
Therefore, the bool has been replaced by a more verbose and type-safe enum
KeyCompare::Mode, where Mode can be either NORMAL (default) or NOCASE.