BugID: 1005
The purpose of this change is to make sure that the UVW vector in the
Measurement Set points from ANTENNA1 to ANTENNA2, that is,

  UVW = UVW(ANTENNA2) - UVW(ANTENNA1)

  In order to comply with WSRT Measurement Sets, it is furthermore
required that

  ANTENNA1 <= ANTENNA2.

  Before this change, the UVW vector pointed exactly in the opposite
direction, hence the UVW vector is reversed. The orientation of the sky was
nevertheless correct. In order to prevent point mirroring the sky with
respect to the phase centre because of the reversal of the UVW coordinates,
it is necessary to conjugate the visibilities prior to writing them to
disk. This conjugation is a temporary fix until the minus signs and
conjugations in the polyphase filter, FFT, fringe tracking,
correlator, and MS exporter are all reviewed in order to determine the proper
place for this conjugation.