Task #3571: Added modificationDate field to OTDBtree information. The field is automatically updated to the current time when a value or limits field in the corresponding tree is updated.
To upgrade an existing database run the 'upgradeOTDB' sql script on the existing database.

A new C++ function and SQL stored procedure is available: getModifiedTrees(after, treetype)
'after' : all trees that where modified after the given time are returned.
'treetype': limit the selection to the given type (0 for no limitation).