Task #8809: MessageBus Py: Resolve PythonControl failure on receiving pipeline feedback msg.
This patch works around the C++ side of MessageBus where XML encoding of special chars
is not done, or perhaps not done if libxml++ is not built.
It'd be better to fix the C++ side than to refrain from encoding on the python side in even
murkier situations, but MessageBus is deprecated and I've spent enough time trying to fix it.

Testing: All MessageBus tests pass on cbt009.